requirements.txt
chromadb
diskcache
gradio-client
groq
ollama
psycopg
llm_axe
selenium
fpdf
docx2pdf
ppt2pdf
fpdf
nara
Pillow
easyocr
numpy
sounddevice
SpeechRecognition
pygame
edge-tts
pyaudio
pyttsx3
colorama
img2pdf
pywin32
opencv-python
pyautogui
beautifulsoup4
pyperclip
pptx
g4f
flask
jinja2